move jira plugin
