+++ packages/cli/package.json
@@ -29,6 +29,7 @@
   "dependencies": {
     "@autorelease/core": "6.5.0",
     "@autorelease/filter-accounts": "6.5.0",
+    "@autorelease/jira": "6.5.0",
     "chalk": "^2.4.2",
     "command-line-args": "^5.1.1",
     "command-line-usage": "^5.0.5",
